requirements.txt
httpx==0.27.0
openai==1.13.3
pandas==2.2.1
Requests==2.31.0
urllib3==1.26.6
Flask[async]==3.0.2
flask-cors==4.0.0
gunicorn==21.2.0
packaging==24.0
gevent==24.2.1
beautifulsoup4==4.12.3
python-dotenv==1.0.1
tavily-python==0.3.3
